pyproject.toml
[tool.poetry]
name = "monocle"
version = "0.1.0"
description = ""
authors = ["widkit <exylloath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
pyside6 = "^6.9.1"
pytablericons = "^1.0.1"
pillow = "^11.2.1"
pytesseract = "^0.3.13"
pyscreenshot = "^3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
